Issue #65 - Add database-backed logging and model updates
Enhanced logging by introducing a custom DbLogger to log messages to the database. Updated `Index.cshtml.cs` to include logging for page load events. Registered `IForumLogService` and its implementation in the DI container.

Modified `DigiozForumContext` to remove default values for certain entity properties. Updated `ForumLang` model to change `LangId` type from `byte` to `short`.

Added `DbLogger` implementation, including `DbLoggerOptions`, `DbLoggerProvider`, and extensions for configuration. Updated `appsettings.json` to include DbLogger settings. Improved maintainability by separating logging logic into dedicated classes and services.

+++ source/digioz.Forum/digioz.Forum/Models/ForumLang.cs
@@ -5,7 +5,7 @@ namespace digioz.Forum.Models;
 
 public partial class ForumLang
 {
-    public byte LangId { get; set; }
+    public short LangId { get; set; }
 
     public string LangIso { get; set; } = null!;
